requirements.txt
# for django
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-environ
# for apps
requests
python-dateutil